# Improved `GMPShortener` performance even further by using `ext-gmp`'s built-in base conversion feature