--- a/src/GMPShortener.php
+++ b/src/GMPShortener.php
@@ -5,17 +5,16 @@ declare(strict_types=1);
 namespace Keiko\Uuid\Shortener;
 
 use Keiko\Uuid\Shortener\Exception\DictionaryException;
-use function gmp_add;
-use function gmp_cmp;
-use function gmp_div_q;
+use function array_flip;
 use function gmp_init;
-use function gmp_intval;
-use function gmp_mod;
 use function gmp_strval;
+use function preg_match;
+use function preg_quote;
 use function str_pad;
 use function str_replace;
+use function strrev;
+use function strtr;
 use function substr;
-use const GMP_ROUND_ZERO;
 use const STR_PAD_LEFT;
 
 /** @psalm-immutable */
@@ -24,41 +23,57 @@ final class GMPShortener extends Shortener
     /** @var Dictionary */
     private $dictionary;
 
+    /**
+     * @var string[] a map of character replacements to be used when translating a number that is
+     *               in base(length(Dictionary)) (keys) to the dictionary characters (values).
+     *
+     * @psalm-var non-empty-array<non-empty-string, non-empty-string>
+     */
+    private $baseReplacements;
+
+    /** @var string */
+    private $allowedCharactersMatcher;
+
     public function __construct(Dictionary $dictionary)
     {
         $this->dictionary = $dictionary;
+
+        $replacements = [];
+
+        foreach (range(0, $dictionary->length - 1) as $characterIndex) {
+            $replacements[gmp_strval(gmp_init((string) $characterIndex, 10), $dictionary->length)] = $dictionary->charsSet[$characterIndex];
+        }
+
+        $this->baseReplacements = $replacements;
+        $this->allowedCharactersMatcher = '/^['. preg_quote($dictionary->charsSet, '/').']+$/';
     }
 
     public function reduce(string $uuid): string
     {
-        $dictionaryLength = gmp_init($this->dictionary->length, 10);
-        $uuidInt = gmp_strval(gmp_init(str_replace('-', '', $uuid), 16));
-        $output = '';
-
-        while (gmp_cmp($uuidInt, '0') > 0) {
-            $output .= $this->dictionary->charsSet[gmp_intval(gmp_mod($uuidInt, $dictionaryLength))];
-
-            $uuidInt = gmp_div_q($uuidInt, $dictionaryLength, GMP_ROUND_ZERO);
-        }
-
-        return $output;
+        return strrev(strtr(
+            gmp_strval(gmp_init(str_replace('-', '', $uuid), 16), $this->dictionary->length),
+            $this->baseReplacements
+        ));
     }
 
     public function expand(string $shortUuid): string
     {
-        $number = gmp_init(0, 10);
-        $dictionaryLength = gmp_init($this->dictionary->length, 10);
-
-        foreach (str_split(strrev($shortUuid)) as $char) {
-            $number = gmp_add(
-                gmp_mul($number, $dictionaryLength),
-                (string) ($this->dictionary->charIndexes[$char] ?? (static function () : string {
-                    throw DictionaryException::indexOutOfBounds();
-                })())
-            );
+        if (1 !== preg_match($this->allowedCharactersMatcher, $shortUuid)) {
+            throw DictionaryException::indexOutOfBounds();
         }
 
-        $base16Uuid = str_pad(gmp_strval($number, 16), 32, '0', STR_PAD_LEFT);
+        $base16Uuid = str_pad(
+            gmp_strval(
+                gmp_init(
+                    strtr(strrev($shortUuid), array_flip($this->baseReplacements)),
+                    $this->dictionary->length
+                ),
+                16
+            ),
+            32,
+            '0',
+            STR_PAD_LEFT
+        );
 
         return substr($base16Uuid,0, 8)
             . '-'