Before:

```
\Benchmark\Keiko\Uuid\Shortener\GMPShortenerBench

    benchShorteningOfTinyUuid...............R5 I4 [μ Mo]/r: 5.160 5.200 (μs) [μSD μRSD]/r: 0.080μs 1.55%
    benchShorteningOfHugeUuid...............R1 I2 [μ Mo]/r: 35.560 35.883 (μs) [μSD μRSD]/r: 0.753μs 2.12%
    benchShorteningOfPromiscuousUuids.......R1 I4 [μ Mo]/r: 3,495.960 3,478.796 (μs) [μSD μRSD]/r: 87.504μs 2.50%
    benchExpandingOfTinyUuid................R5 I4 [μ Mo]/r: 7.160 7.122 (μs) [μSD μRSD]/r: 0.150μs 2.09%
    benchExpandingOfHugeUuid................R1 I0 [μ Mo]/r: 25.920 26.211 (μs) [μSD μRSD]/r: 0.627μs 2.42%
    benchExpandingOfPromiscuousUuids........R2 I1 [μ Mo]/r: 2,442.840 2,425.225 (μs) [μSD μRSD]/r: 53.328μs 2.18%
```

After:

```
\Benchmark\Keiko\Uuid\Shortener\GMPShortenerBench

    benchShorteningOfTinyUuid...............R1 I0 [μ Mo]/r: 4.000 4.000 (μs) [μSD μRSD]/r: 0.000μs 0.00%
    benchShorteningOfHugeUuid...............R1 I4 [μ Mo]/r: 4.200 4.085 (μs) [μSD μRSD]/r: 0.179μs 4.26%
    benchShorteningOfPromiscuousUuids.......R1 I2 [μ Mo]/r: 321.840 325.473 (μs) [μSD μRSD]/r: 6.486μs 2.02%
    benchExpandingOfTinyUuid................R5 I4 [μ Mo]/r: 19.240 19.649 (μs) [μSD μRSD]/r: 0.637μs 3.31%
    benchExpandingOfHugeUuid................R5 I4 [μ Mo]/r: 19.640 19.423 (μs) [μSD μRSD]/r: 0.427μs 2.17%
    benchExpandingOfPromiscuousUuids........R3 I3 [μ Mo]/r: 609.600 617.942 (μs) [μSD μRSD]/r: 16.369μs 2.69%
```

This is another 10x improvement on Promiscuous UUID shortening, which leads to 100x improvement over previous/baseline functionality